LPS-172073 search-experiences-web: Separate modals into own component
