+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/query_builder_tab/SelectTypes.js
@@ -10,16 +10,12 @@
  */
 
 import ClayButton from '@clayui/button';
-import ClayEmptyState from '@clayui/empty-state';
-import {ClayCheckbox} from '@clayui/form';
 import ClayIcon from '@clayui/icon';
-import ClayModal, {useModal} from '@clayui/modal';
 import ClayTable from '@clayui/table';
-import {ManagementToolbar} from 'frontend-js-components-web';
 import React, {useContext, useState} from 'react';
 
 import ThemeContext from '../../shared/ThemeContext';
-import sub from '../../utils/language/sub';
+import SearchableTypesModal from './SearchableTypesModal';
 
 function SelectTypes({
 	onFrameworkConfigChange,
@@ -30,43 +26,17 @@ function SelectTypes({
 	const {locale} = useContext(ThemeContext);
 
 	const [visible, setVisible] = useState(false);
-	const {observer, onClose} = useModal({
-		onClose: () => setVisible(false),
-	});
-	const [modalSelectedTypes, setModalSelectedTypes] = useState(selectedTypes);
 
-	const searchableTypesClassNames = searchableTypes.map(
-		({className}) => className
-	);
 	const searchableTypesSorted = searchableTypes.sort((a, b) =>
 		a.displayName.localeCompare(b.displayName, locale.replace('_', '-'))
 	);
 
 	const _handleDelete = (type) => () => {
-		const newSelected = modalSelectedTypes.filter((item) => item !== type);
+		const newSelected = selectedTypes.filter((item) => item !== type);
 
 		onFrameworkConfigChange({
 			searchableAssetTypes: newSelected,
 		});
-		setModalSelectedTypes(newSelected);
-	};
-
-	const _handleModalDone = () => {
-		onClose();
-
-		onFrameworkConfigChange({
-			searchableAssetTypes: modalSelectedTypes,
-		});
-	};
-
-	const _handleRowCheck = (type) => () => {
-		const isSelected = modalSelectedTypes.includes(type);
-
-		setModalSelectedTypes(
-			isSelected
-				? modalSelectedTypes.filter((item) => item !== type)
-				: [...modalSelectedTypes, type]
-		);
 	};
 
 	return (
@@ -76,7 +46,6 @@ function SelectTypes({
 				displayType="secondary"
 				onClick={() => {
 					setVisible(true);
-					setModalSelectedTypes(selectedTypes);
 				}}
 				small
 			>
@@ -116,170 +85,13 @@ function SelectTypes({
 			)}
 
 			{visible && (
-				<ClayModal
-					className="modal-height-xl sxp-searchable-types-modal-root"
-					observer={observer}
-					size="lg"
-				>
-					<ClayModal.Header>
-						{Liferay.Language.get('select-types')}
-					</ClayModal.Header>
-
-					{searchableTypes.length ? (
-						<>
-							<ManagementToolbar.Container
-								className={
-									!!modalSelectedTypes.length &&
-									'management-bar-primary'
-								}
-							>
-								<div className="navbar-form navbar-form-autofit navbar-overlay">
-									<ManagementToolbar.ItemList>
-										<ManagementToolbar.Item>
-											<ClayCheckbox
-												checked={
-													!!modalSelectedTypes.length
-												}
-												indeterminate={
-													!!modalSelectedTypes.length &&
-													modalSelectedTypes.length !==
-														searchableTypes.length
-												}
-												onChange={() =>
-													setModalSelectedTypes(
-														!modalSelectedTypes.length
-															? searchableTypesClassNames
-															: []
-													)
-												}
-											/>
-										</ManagementToolbar.Item>
-
-										<ManagementToolbar.Item>
-											{modalSelectedTypes.length ? (
-												<>
-													<span className="component-text">
-														{sub(
-															Liferay.Language.get(
-																'x-of-x-selected'
-															),
-															[
-																modalSelectedTypes.length,
-																searchableTypes.length,
-															],
-															false
-														)}
-													</span>
-
-													{modalSelectedTypes.length <
-														searchableTypes.length && (
-														<ClayButton
-															displayType="link"
-															onClick={() => {
-																setModalSelectedTypes(
-																	searchableTypesClassNames
-																);
-															}}
-															small
-														>
-															{Liferay.Language.get(
-																'select-all'
-															)}
-														</ClayButton>
-													)}
-												</>
-											) : (
-												<span className="component-text">
-													{Liferay.Language.get(
-														'select-all'
-													)}
-												</span>
-											)}
-										</ManagementToolbar.Item>
-									</ManagementToolbar.ItemList>
-								</div>
-							</ManagementToolbar.Container>
-
-							<ClayModal.Body scrollable>
-								<ClayTable>
-									<ClayTable.Body>
-										{searchableTypesSorted.map(
-											({className, displayName}) => {
-												const isSelected = modalSelectedTypes.includes(
-													className
-												);
-
-												return (
-													<ClayTable.Row
-														active={isSelected}
-														key={className}
-														onClick={_handleRowCheck(
-															className
-														)}
-													>
-														<ClayTable.Cell>
-															<ClayCheckbox
-																checked={
-																	isSelected
-																}
-																onChange={_handleRowCheck(
-																	className
-																)}
-															/>
-														</ClayTable.Cell>
-
-														<ClayTable.Cell
-															expanded
-															headingTitle
-														>
-															{displayName}
-														</ClayTable.Cell>
-													</ClayTable.Row>
-												);
-											}
-										)}
-									</ClayTable.Body>
-								</ClayTable>
-							</ClayModal.Body>
-						</>
-					) : (
-						<ClayModal.Body>
-							<ClayEmptyState
-								description={Liferay.Language.get(
-									'an-error-has-occurred-and-we-were-unable-to-load-the-results'
-								)}
-								imgSrc="/o/admin-theme/images/states/empty_state.gif"
-								title={Liferay.Language.get(
-									'no-items-were-found'
-								)}
-							>
-								<ClayButton
-									displayType="secondary"
-									onClick={onFetchSearchableTypes}
-								>
-									{Liferay.Language.get('refresh')}
-								</ClayButton>
-							</ClayEmptyState>
-						</ClayModal.Body>
-					)}
-
-					<ClayModal.Footer
-						last={
-							<ClayButton.Group spaced>
-								<ClayButton
-									displayType="secondary"
-									onClick={onClose}
-								>
-									{Liferay.Language.get('cancel')}
-								</ClayButton>
-
-								<ClayButton onClick={_handleModalDone}>
-									{Liferay.Language.get('done')}
-								</ClayButton>
-							</ClayButton.Group>
-						}
-					/>
-				</ClayModal>
+				<SearchableTypesModal
+					onFetchSearchableTypes={onFetchSearchableTypes}
+					onFrameworkConfigChange={onFrameworkConfigChange}
+					searchableTypes={searchableTypesSorted}
+					selectedTypes={selectedTypes}
+					setVisible={setVisible}
+				/>
 			)}
 		</>
 	);
